requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
